+++ lib/types.ts
@@ -209,8 +209,6 @@ export type DomainCheckResult = {
   premium: boolean;
 };
 
-export type DomainPremiumPrice = { premium_price: string; action: string };
-
 export type DomainPrices = {
   domain: string;
   premium: boolean;
@@ -270,17 +268,6 @@ export type WhoisPrivacy = {
   updated_at: string;
 };
 
-export type WhoisPrivacyRenewal = {
-  id: number;
-  domain_id: number;
-  whois_privacy_id: number;
-  state: string;
-  enabled: boolean;
-  expires_on: string;
-  created_at: string;
-  updated_at: string;
-};
-
 export type PrimaryServer = {
   id: number;
   account_id: number;
@@ -308,27 +295,6 @@ export type ZoneFile = { zone: string };
 
 export type ZoneDistribution = { distributed: boolean };
 
-export type ZoneRecordType =
-  | "A"
-  | "AAAA"
-  | "ALIAS"
-  | "CAA"
-  | "CNAME"
-  | "DNSKEY"
-  | "DS"
-  | "HINFO"
-  | "MX"
-  | "NAPTR"
-  | "NS"
-  | "POOL"
-  | "PTR"
-  | "SOA"
-  | "SPF"
-  | "SRV"
-  | "SSHFP"
-  | "TXT"
-  | "URL";
-
 export type ZoneRecordRegion =
   | "global"
   | "SV1"
@@ -348,7 +314,7 @@ export type ZoneRecord = {
   content: string;
   ttl: number;
   priority?: number | null;
-  type: ZoneRecordType;
+  type: string;
   regions: Array<ZoneRecordRegion>;
   system_record: boolean;
   created_at: string;
@@ -408,27 +374,6 @@ export type Template = {
   updated_at: string;
 };
 
-export type TemplateRecordType =
-  | "A"
-  | "AAAA"
-  | "ALIAS"
-  | "CAA"
-  | "CNAME"
-  | "DNSKEY"
-  | "DS"
-  | "HINFO"
-  | "MX"
-  | "NAPTR"
-  | "NS"
-  | "POOL"
-  | "PTR"
-  | "SOA"
-  | "SPF"
-  | "SRV"
-  | "SSHFP"
-  | "TXT"
-  | "URL";
-
 export type TemplateRecord = {
   id: number;
   template_id: number;
@@ -436,7 +381,7 @@ export type TemplateRecord = {
   content: string;
   ttl: number;
   priority: number | null;
-  type: TemplateRecordType;
+  type: string;
   created_at: string;
   updated_at: string;
 };